pyproject.toml
[project]
name = "pdfllm"
version = "0.2.0"
description = "A股并购报告PDF结构化事实抽取系统"
requires-python = ">=3.11"
dependencies = [
    "pymupdf4llm",
    "pymupdf",
    "sentence-transformers",
    "faiss-cpu",
    "openai",
    "pydantic>=2.0",
    "pydantic-settings",
    "python-dotenv",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[project.scripts]
pdfllm = "scripts.run:main"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
